feat: Add getParticipants route
The code changes add a new route called `getParticipants` to retrieve participants for a specific trip. This route is implemented in the `get-participants.ts` file. The changes also include importing necessary types and dependencies, as well as handling errors with the `ClientError` class.

Note: The suggested commit message follows the conventional format of "feat: <commit message>".